+++ llvm/unittests/Support/RISCVISAInfoTest.cpp
@@ -762,6 +762,10 @@ Experimental extensions
     zicfilp             0.2       This is a long dummy description
     zicond              1.0
     zacas               1.0
+<<<<<<< HEAD
+=======
+    zfa                 1.0
+>>>>>>> origin/sifive-dev
     zfbfmin             0.8
     ztso                0.1
     zvbb                1.0